fix float -> int warning

+++ src/color.cc
@@ -349,7 +349,7 @@ rgba_from_rgba_string(const char *str, short *ok) {
         }
       }
     }
-    return *ok = 1, rgba_from_rgba(r, g, b, a * 255);
+    return *ok = 1, rgba_from_rgba(r, g, b, (int) (a * 255));
   }
   return *ok = 0;
 }
